+++ resources/js/app.js
@@ -1,3 +1,4 @@
 import './bootstrap';
 import './pwa-shell';
+import './public-connectivity';
 import './public-mobile-navigation';